pinyintools: use regex for final matching instead
The use of str.rsplit() to tokenize the initial and final would result
in a null initial if the initial was a substring of the final.

For example, 南 (nán) would tokenize to ["", "an", 2] instead of the
expected ["n", "an", 2] because the initial "n" is a substring of the
final "an".